pyproject.toml
[project]
name = "openai-sdk-test"
version = "0.1.0"
description = ""
authors = [
    {name = "sial-sanwal",email = "sialsanwal885@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents (>=0.0.4,<0.0.5)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
